+++ packages/react-native/src/pages/Gamification/Gamification.tsx
@@ -1,25 +1,14 @@
-import { useEffect, useState } from 'react';
 import QuizLoading from '@components/gamification/QuizLoading';
 import QuizSlider from '@components/gamification/QuizSlider';
 import withSuspense from '@/components/HOC/withSuspense';
-import useGeolocation from '@/hooks/useGeolocation';
-import useQuizzesQuery, { Location } from '@/apis/queries/quiz/useQuizzesQuery';
+import useQuizzesQuery from '@/apis/queries/quiz/useQuizzesQuery';
 import NoQuiz from '@/components/gamification/NoQuiz';
+import useLocation from '@/hooks/useLocation';
 
 export default withSuspense(
   function Gamification() {
-    const { getGeolocation } = useGeolocation();
-    const [location, setLocation] = useState<Location>();
+    const { location } = useLocation();
     const { data } = useQuizzesQuery({ location });
-
-    useEffect(() => {
-      getGeolocation().then((res) => {
-        setLocation({
-          latitude: res?.coords.latitude,
-          longitude: res?.coords.longitude,
-        });
-      });
-    }, []);
 
     if (!location?.latitude || !location.longitude) {
       return <QuizLoading />;
